Fix gpg already running error on MacOS
I had commented out starting gpg-agent on teminal session because it
kept giving me an annoying error message. To work around I just manually
called the eval command whenever I needed to use gpg and it would run
for the remainder of my session.

Now I'm checking if gpg-agent is already running and start it if the
/bye argument exit code returns an error. It won't fix all use cases,
but hopefully the more frequent ones.