feat: Add install script for Pop_OS
This adds a bash install script for pop_os, I'd like to expand this to
work on WSL and macOS at the least.

I'd like to eventually phase out dotbot because I'm not a huge fan of
having a python dependency. It'd be nice to have a single binary that
performs the same functionality of dotbot.